Kafka_Connect/requirements.txt
pandas
numpy
requests
kafka-python
mysql-connector-python